Created item card and layout of first screen using temp data

+++ app/src/main/java/com/phoenix/appjumpstart/data/model/Item.kt
@@ -2,6 +2,6 @@ package com.phoenix.appjumpstart.data.model
 
 data class Item(
     val name:String = "",
-    val price:Int = 0,
+    val price:Double = 0.0,
     val isSameDayShipping: Boolean = false
 )
